chore: migrate rust/parallel_calls to icp-cli
Replace dfx.json with icp.yaml (@dfinity/rust@v3.3.0), move sources
from src/caller/ and src/callee/ to top-level caller/ and callee/
subdirs, update caller to discover callee via PUBLIC_CANISTER_ID:callee
env var (eliminating setup_callee), upgrade ic-cdk to 0.20 with the
new Call::bounded_wait API, add Makefile with numbered tests for
sequential/parallel behaviour, and add rust-parallel_calls jobs to the
existing parallel_calls.yml CI workflow.
