+++ rust/parallel_calls/Makefile
@@ -1,21 +1,37 @@
-.PHONY: all
-all: test
+.PHONY: test test-multi-subnet
 
-.PHONY: deploy
-.SILENT: deploy
-build:
-	dfx deploy --no-wallet
+test:
+	@echo "=== Test 1: sequential_calls(10) — all 10 succeed ==="
+	@result=$$(icp canister call caller sequential_calls '(10)') && \
+	  echo "$$result" && \
+	  echo "$$result" | grep -q '(10 : nat64)' && \
+	  echo "PASS" || (echo "FAIL" && exit 1)
 
-.PHONY: test
-.SILENT: test
-test: build
-	dfx canister call caller setup_callee "(principal \"`dfx canister id callee`\")"
-	dfx canister call caller sequential_calls 100 | grep '(100 : nat64)' && echo 'PASS'
-	dfx canister call caller parallel_calls 100 | grep '(100 : nat64)' && echo 'PASS'
-	dfx canister call caller sequential_calls 2000 | grep '(2_000 : nat64)' && echo 'PASS' 
-	dfx canister call caller parallel_calls 2000 | grep '(500 : nat64)' && echo 'PASS'
+	@echo "=== Test 2: parallel_calls(10) — all 10 succeed (same result as sequential at low n) ==="
+	@result=$$(icp canister call caller parallel_calls '(10)') && \
+	  echo "$$result" && \
+	  echo "$$result" | grep -q '(10 : nat64)' && \
+	  echo "PASS" || (echo "FAIL" && exit 1)
 
-.PHONY: clean
-.SILENT: clean
-clean:
-	rm -rf .dfx
+	@echo "=== Test 3: sequential_calls(2000) — all succeed (one in-flight at a time) ==="
+	@result=$$(icp canister call caller sequential_calls '(2000)') && \
+	  echo "$$result" && \
+	  echo "$$result" | grep -q '(2_000 : nat64)' && \
+	  echo "PASS" || (echo "FAIL" && exit 1)
+
+	@echo "=== Test 4: parallel_calls(2000) — fewer succeed (in-flight limit exceeded) ==="
+	@# The replica limits in-flight calls per canister, so most parallel calls fail at n=2000.
+	@# Extract the number and assert it is strictly less than 2000.
+	@result=$$(icp canister call caller parallel_calls '(2000)') && \
+	  echo "$$result" && \
+	  n=$$(echo "$$result" | grep -oE '[0-9_]+' | head -1 | tr -d '_') && \
+	  [ "$$n" -lt 2000 ] && \
+	  echo "PASS" || (echo "FAIL" && exit 1)
+
+test-multi-subnet:
+	@echo "=== Building Rust WASMs ==="
+	cargo build --target wasm32-unknown-unknown --release -p caller -p callee
+	@echo "=== Running multi-subnet PocketIC test ==="
+	CALLER_WASM=target/wasm32-unknown-unknown/release/caller.wasm \
+	CALLEE_WASM=target/wasm32-unknown-unknown/release/callee.wasm \
+	  cargo run --manifest-path multi_subnet/Cargo.toml
