+++ rust/parallel_calls/caller/Cargo.toml
@@ -3,12 +3,10 @@ name = "caller"
 version = "0.1.0"
 edition = "2021"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
 [lib]
 crate-type = ["cdylib"]
 
 [dependencies]
 candid = "0.10"
-ic-cdk = "0.17"
+ic-cdk = "0.20"
 futures = "0.3"
