--- rust/parallel_calls/dfx.json
@@ -1,22 +1,0 @@
-{
-  "canisters": {
-    "caller": {
-      "candid": "src/caller/caller.did",
-      "package": "caller",
-      "type": "rust"
-    },
-    "callee": {
-      "candid": "src/callee/callee.did",
-      "package": "callee",
-      "type": "rust"
-    }
-  },
-  "defaults": {
-    "build": {
-      "args": "",
-      "packtool": ""
-    }
-  },
-  "output_env_file": ".env",
-  "version": 1
-}
