+++ rust/parallel_calls/icp.yaml
@@ -1,0 +1,11 @@
+networks:
+  - name: local
+    mode: managed
+canisters:
+  - name: caller
+    recipe:
+      type: "@dfinity/rust@v3.3.0"
+
+  - name: callee
+    recipe:
+      type: "@dfinity/rust@v3.3.0"
