+++ rust/parallel_calls/multi_subnet/src/main.rs
