+++ rust/parallel_calls/rust-toolchain.toml
@@ -1,3 +1,2 @@
 [toolchain]
-channel = "1.78.0"
 targets = ["wasm32-unknown-unknown"]
