--- rust/parallel_calls/src/callee/callee.did
@@ -1,1 +1,0 @@
-service : { ping : () -> () }
