--- rust/parallel_calls/src/caller/caller.did
@@ -1,5 +1,0 @@
-service : {
-  parallel_calls : (nat64) -> (nat64);
-  sequential_calls : (nat64) -> (nat64);
-  setup_callee : (principal) -> ();
-}
